De-centralize onBackspaceResetFormattingAtDocumentStart behavior
