+++ packages/slate-editor/src/extensions/blockquote/BlockquoteExtension.tsx
@@ -2,7 +2,7 @@ import type { Extension } from '@prezly/slate-commons';
 import { isQuoteNode, QUOTE_NODE_TYPE } from '@prezly/slate-types';
 import React from 'react';
 
-import { withResetFormattingOnBreak } from '#lib';
+import {onBackspaceResetFormattingAtDocumentStart, withResetFormattingOnBreak} from '#lib';
 
 import { BlockQuoteElement } from './components';
 import { normalizeRedundantAttributes } from './lib';
@@ -19,6 +19,9 @@ export function BlockquoteExtension(): Extension {
             },
         },
         normalizeNode: [normalizeRedundantAttributes],
+        onKeyDown(event, editor) {
+            onBackspaceResetFormattingAtDocumentStart(editor, isQuoteNode, event);
+        },
         renderElement({ attributes, element, children }) {
             if (isQuoteNode(element)) {
                 return (
