+++ packages/slate-editor/src/extensions/rich-formatting/RichFormattingExtension.tsx
@@ -15,7 +15,6 @@ export const RichFormattingExtension = (): Extension => ({
     normalizeNode: normalizeRedundantRichTextAttributes,
     onKeyDown: (event: KeyboardEvent, editor: Editor) => {
         OnKeyDown.onHotkeyDoMarks(event, editor);
-        OnKeyDown.onBackspaceResetFormattingAtDocumentStart(event, editor);
     },
     renderLeaf: Text,
 });
