+++ packages/slate-editor/src/extensions/rich-formatting/index.ts
@@ -1,6 +1,6 @@
 export { RichFormattingExtension, EXTENSION_ID } from './RichFormattingExtension';
 
-export { RichTextElement, Text } from './components';
+export { Text } from './components';
 export { isRichTextBlockElement, isRichTextElement, toggleBlock } from './lib';
 export type { RichFormattingExtensionParameters, RichTextElementType } from './types';
 export { ElementType, MarkType } from './types';
