+++ packages/slate-editor/src/extensions/rich-formatting/onKeyDown.ts
@@ -1,11 +1,7 @@
 import { EditorCommands } from '@prezly/slate-commons';
-import { isHeadingNode, isQuoteNode, PARAGRAPH_NODE_TYPE } from '@prezly/slate-types';
 import { isHotkey } from 'is-hotkey';
 import type { KeyboardEvent } from 'react';
-import type { Node, Path, Point } from 'slate';
-import { Editor, Range, Transforms } from 'slate';
-
-import { isDeletingEventBackward } from '#lib';
+import type { Editor } from 'slate';
 
 import { MarkType } from './types';
 
@@ -22,41 +18,4 @@ export function onHotkeyDoMarks(event: KeyboardEvent, editor: Editor) {
             EditorCommands.toggleMark(editor, mark);
         }
     });
-}
-
-export function onBackspaceResetFormattingAtDocumentStart(event: KeyboardEvent, editor: Editor) {
-    const { selection } = editor;
-    if (
-        isDeletingEventBackward(event) &&
-        selection !== null &&
-        Range.isCollapsed(selection) &&
-        isDocumentStart(selection.focus)
-    ) {
-        if (isFocused(editor, isRichBlock)) {
-            Transforms.setNodes(
-                editor,
-                { type: PARAGRAPH_NODE_TYPE },
-                {
-                    match: (node, path) =>
-                        isRichBlock(node) && EditorCommands.isTopLevelNode(node, path),
-                },
-            );
-        }
-    }
-}
-
-function isDocumentStart(focus: Point): boolean {
-    return (
-        focus.offset === 0 &&
-        // Any node at path [0, ..., 0] is considered the start of the document
-        focus.path.every((index) => index === 0)
-    );
-}
-
-function isFocused(editor: Editor, match: (node: Node, path: Path) => boolean): boolean {
-    return Array.from(Editor.nodes(editor, { match })).length > 0;
-}
-
-function isRichBlock(node: Node) {
-    return isQuoteNode(node) || isHeadingNode(node);
 }
